ft_strlen: int -> size_t. Correct

+++ Rank00/Libft/ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: analmeid <analmeid@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/10 12:06:19 by analmeid          #+#    #+#             */
-/*   Updated: 2024/04/10 12:06:43 by analmeid         ###   ########.fr       */
+/*   Updated: 2024/06/18 16:30:22 by analmeid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 // calculate the length of a string
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
